Moving visual tree extensions out and limiting the search for finding element in error
