+++ src/ServiceControl.Config/UI/InstanceAdd/InstanceAddAttachment.cs
@@ -48,7 +48,7 @@
             {
                 viewModel.NotifyOfPropertyChange(string.Empty);
                 viewModel.SubmitAttempted = false;
-                windowManager.ScrollFirstErrorIntoView();
+                windowManager.ScrollFirstErrorIntoView(viewModel);
                 
                 return;
             }
